Fix the new warnings.

+++ crawl-ref/source/output.cc
@@ -2426,7 +2426,7 @@ static std::string _status_mut_abilities(int sw)
     {
         text += make_stringf("\n<w>%s:</w> %d/%d rune%s: %s",
                     stringize_glyph(get_item_symbol(SHOW_ITEM_MISCELLANY)).c_str(),
-                    runes.size(), you.obtainable_runes,
+                    (int)runes.size(), you.obtainable_runes,
                     you.obtainable_runes == 1 ? "" : "s",
                     comma_separated_line(runes.begin(), runes.end(),
                                          ", ", ", ").c_str());
